+++ src/backend/joanie/tests/core/test_api_certificate.py
@@ -52,7 +52,7 @@ class CertificateApiTest(BaseAPITestCase):
 
         token = self.generate_token_from_user(user)
 
-        with self.assertNumQueries(2):
+        with self.assertNumQueries(4):
             response = self.client.get(
                 "/api/v1.0/certificates/", HTTP_AUTHORIZATION=f"Bearer {token}"
             )
@@ -90,6 +90,8 @@ class CertificateApiTest(BaseAPITestCase):
                                 "logo": "_this_field_is_mocked",
                                 "title": order.organization.title,
                             },
+                            "learner_name": certificate.order.owner.get_full_name(),
+                            "title": certificate.order.product.title,
                         },
                     },
                 ],
@@ -216,6 +218,8 @@ class CertificateApiTest(BaseAPITestCase):
                         "logo": "_this_field_is_mocked",
                         "title": certificate.order.organization.title,
                     },
+                    "learner_name": certificate.order.owner.get_full_name(),
+                    "title": certificate.order.product.title,
                 },
             },
         )